backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
numpy==1.26.4
PyPDF2==3.0.1
python-docx==1.1.0
python-magic==0.4.27
requests==2.31.0
pinecone==7.3.0
python-dotenv==1.0.0
